+++ CHANGELOG.md
@@ -3,11 +3,13 @@ Changes
 
 1.0.0 to 1.1.0
 --------------
+* [UI]: Samples in the cart are now sorted by created date.
 * [UI]: Fixed a bug where exporting to Galaxy from the cart was not working.
 * [Database] Fixed an issue with migrating single end sequence files to the new SequencingObject model in cases where samples with single files had been merged.  IRIDA installs should skip directly to v1.0.2 from 1.0.0alpha-10 to avoid database update problems. (1.0.2)
 * [Developer]: Added a complete docker image for Galaxy, updated the VirtualBox appliance to use the Docker image.
 * [UI]: Fixed a bug when the launch pipeline page was opened on a small display (<1000px) the "Launch Pipeline" button disappeared.
 * [Developer]: Added support for a wider variety of naming patters when pairing uploaded sequence files. 
+* [UI]: Fixed a bug where the `hashCode` method on `SequenceFilePair` was *only* using the date the pair was created, causing `Set`s of pairs to be much smaller than expected. (1.0.3)
 
 1.0.0-alpha10 to 1.0.0
 ----------------------